chore: cleanups, extract common utils for various date pickers
